backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
asyncpg
python-dotenv
httpx
pydantic
openai
google-generativeai
